python2

+++ lib/yolo.py
@@ -100,7 +100,7 @@ class YOLO(object):
                     dtype = '({})float32,'.format(bias_number))
                 offset += bias_number * 4
             
-                chunks  = [chunk] * (weight_number / chunk) 
+                chunks  = [chunk] * int(weight_number / chunk) 
                 chunks += [weight_number % chunk]
                 l.weights = np.array([], dtype = np.float32)
                 for c in chunks:
